chore: main의 컴포넌트 testpage로 이동
